Extract POSItemIdentifier to its own file
This prevents the generated code from importing PagedItems from networking.
